fix: Change the name and move the folder
